User Service - basic operations; Data Service - refactoring

+++ device-service/src/main/java/org/smartlights/device/resources/DeviceResource.java
@@ -24,5 +24,5 @@ public interface DeviceResource {
 
     @GET
     @Path("/data/count")
-    int getCountOfData();
+    Long getCountOfData();
 }
